lambdas/chalice/pull-google-news/requirements.txt
setuptools<58
pygooglenews
chalice
newspaper3k
psycopg2-binary
